% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_url.R
\name{create_shinylive_url}
\alias{create_shinylive_url}
\title{Creates Shinylive url for the app code.}
\usage{
create_shinylive_url(code)
}
\arguments{
\item{code}{(\code{character(1)}) A string with app code.}
}
\value{
(\code{character(1)}) Shinylive app url.
}
\description{
Creates Shinylive url for the app code.
}
\examples{
code <- "this is your app code as a string"
create_shinylive_url(code)
}
