% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesShinyLive.R
\name{tag-examplesShinyLive}
\alias{tag-examplesShinyLive}
\title{Custom \code{examplesShinyLive} tag.}
\usage{
#' @examplesShinyLive${1:# example code (optional)}
}
\description{
This function generates a new section with ShinyLive links to the applications from this tag content.
If no code is provided then the code from the following \verb{@examples} tag is used.
}
\details{
The application code must be executable inside ShinyLive. If the application code includes functions from your
package, you must add \verb{library(<your package>)} beforehand. For more information, refer to the Decoration section
on how to use and decorate existing examples.

Note: All the packages used in the application code need to be installable in WebR.
See \href{https://docs.r-wasm.org/webr/latest/packages.html}{this article} for more details.
}
\section{Decoration}{


To avoid repetition between the \verb{@examplesShinyLive} and \verb{@examples} sections, there are special string literals
that provide access to the \verb{@examples} content from within \verb{@examplesShinyLive}.
These literals should be used as expressions embraced with \code{{{ }}}, which are then interpolated using
\code{glue::glue_data(..., .open = "{{", .close = "}}")}.

The following keywords are available:
\itemize{
\item \code{"{{ tags_examples }}"} - a list of tags with examples
\item \code{"{{ examples }}"} - a list of "raw" elements from \code{tags_examples}
\item \code{"{{ next_example }}"} - "raw" element of the next example
\item \code{"{{ prev_example }}"} - "raw" element of the previous example
}

This allows you to access and decorate existing example code to create executable application code for ShinyLive.
Refer to the examples section for possible use cases.
}

\examples{
# As a part of documentation:

# basic example:
#' (docs)
#' @examplesShinyLive
#' @examples
#' (example code)

# using keywords:
#' (docs)
#' @examplesShinyLive
#' foo <- 1
#' {{ next_example }}
#' bar <- 2
#' @examples
#' (your example code)

# A typical example would be:
#' (docs)
#' @examplesShinyLive
#' library(<your package>)
#' interactive <- function() TRUE
#' {{ next_example }}
#' @examples
#' app <- ...
#' if (interactive()) {
#'   shinyApp(app$ui, app$server)
#' }

# multiple apps:
#' (docs)
#' @examplesShinyLive
#' @examples
#' (your example app 1)
#' @examplesShinyLive
#' @examples
#' (your example app 2)

# skip parts of example code:
#' (docs)
#' @examples
#' (your example code - skipped)
#' @examplesShinyLive
#' @examples
#' (your example code - included)

# multiple apps with keywords:
#' (docs)
#' @examplesShinyLive
#' x <- 1
#' {{ next_example }}
#' @examples
#' (your example app 1)
#' @examplesShinyLive
#' y <- 1
#' {{ next_example }}
#' @examples
#' (your example app 2)
}
